requirements-test.txt
pynose==1.5.3
coverage==7.6.2
coveralls==4.0.1
responses==0.25.3
ruff==0.6.9
